Endpoint: Move files, add README, replace implicit `any` with stricter types. Reorganize types.  (#63262)
* Removed `FIXME` comments and references to private repos. Please use Github issues to track work
* Added a README describing the modules in `applications/endpoint`
* Removed dead code
* Moved `AppRoot` component to its own module
* Moved `applications/endpoint/services` under `store`
* Moved some exported types to `applications/endpoint/types`
* Moved all React code to `view`
* Added types in some places that were implicitly `any`
* Moved `PageId` type from common directory (where it could be shared with the server) to the public directory

Conflicts:
* `x-pack/plugins/endpoint/public/applications/endpoint/view/policy/policy_details.tsx`
