+++ x-pack/plugins/endpoint/common/generate_data.ts
@@ -7,7 +7,6 @@
 import uuid from 'uuid';
 import seedrandom from 'seedrandom';
 import { AlertEvent, EndpointEvent, HostMetadata, OSFields, HostFields } from './types';
-// FIXME: move types/model to top-level
 // eslint-disable-next-line @kbn/eslint/no-restricted-paths
 import { PolicyData } from '../public/applications/endpoint/types';
 // eslint-disable-next-line @kbn/eslint/no-restricted-paths
