+++ x-pack/plugins/endpoint/public/applications/endpoint/view/components/link_to_app.test.tsx
@@ -7,9 +7,14 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import { LinkToApp } from './link_to_app';
-import { KibanaContextProvider } from '../../../../../../../src/plugins/kibana_react/public';
-import { coreMock } from '../../../../../../../src/core/public/mocks';
 import { CoreStart } from 'kibana/public';
+import { KibanaContextProvider } from '../../../../../../../../src/plugins/kibana_react/public';
+import { coreMock } from '../../../../../../../../src/core/public/mocks';
+
+type LinkToAppOnClickMock<Return = void> = jest.Mock<
+  Return,
+  [React.MouseEvent<HTMLAnchorElement, MouseEvent>]
+>;
 
 describe('LinkToApp component', () => {
   let fakeCoreStart: jest.Mocked<CoreStart>;
@@ -38,7 +43,8 @@ describe('LinkToApp component', () => {
     ).toMatchSnapshot();
   });
   it('should support onClick prop', () => {
-    const spyOnClickHandler = jest.fn();
+    // Take `_event` (even though it is not used) so that `jest.fn` will have a type that expects to be called with an event
+    const spyOnClickHandler: LinkToAppOnClickMock = jest.fn(_event => {});
     const renderResult = render(
       <LinkToApp appId="ingestManager" href="/app/ingest" onClick={spyOnClickHandler}>
         link
@@ -91,7 +97,8 @@ describe('LinkToApp component', () => {
     });
   });
   it('should still preventDefault if onClick callback throws', () => {
-    const spyOnClickHandler = jest.fn(ev => {
+    // Take `_event` (even though it is not used) so that `jest.fn` will have a type that expects to be called with an event
+    const spyOnClickHandler: LinkToAppOnClickMock<never> = jest.fn(_event => {
       throw new Error('test');
     });
     const renderResult = render(
@@ -104,7 +111,7 @@ describe('LinkToApp component', () => {
     expect(clickEventArg.isDefaultPrevented()).toBe(true);
   });
   it('should not navigate if onClick callback prevents defalut', () => {
-    const spyOnClickHandler = jest.fn(ev => {
+    const spyOnClickHandler: LinkToAppOnClickMock = jest.fn(ev => {
       ev.preventDefault();
     });
     const renderResult = render(
