+++ x-pack/plugins/endpoint/public/applications/endpoint/view/components/page_view.tsx
