+++ x-pack/plugins/endpoint/public/applications/endpoint/view/policy/policy_list.tsx
@@ -23,8 +23,8 @@ import { usePolicyListSelector } from './policy_hooks';
 import { PolicyListAction } from '../../store/policy_list';
 import { PolicyData } from '../../types';
 import { useKibana } from '../../../../../../../../src/plugins/kibana_react/public';
-import { PageView } from '../../components/page_view';
-import { LinkToApp } from '../../components/link_to_app';
+import { PageView } from '../components/page_view';
+import { LinkToApp } from '../components/link_to_app';
 
 interface TableChangeCallbackArguments {
   page: { index: number; size: number };
